requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.34.90
botocore==1.34.90
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.15.1
html-sanitizer==2.4.1
idna==3.7
jmespath==1.0.1
jsonfield==3.1.0
lxml==5.2.1
lxml_html_clean==0.1.1
pillow==10.3.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
sendgrid==6.6.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
stripe==9.3.0
typing_extensions==4.11.0
urllib3==2.2.1
uWSGI==2.0.25.1
